requirements.txt
ultralytics
opencv-python
Pillow
imagehash
icrawler
scikit-learn
numpy<2
opencv-python
